chore(webhooks): Rename webhooks application to webhook_listeners
There is already a webhooks application in the appstore
