+++ .gitignore
@@ -45,7 +45,7 @@
 !/apps/twofactor_backupcodes
 !/apps/user_status
 !/apps/weather_status
-!/apps/webhooks
+!/apps/webhook_listeners
 !/apps/workflowengine
 /apps/files_external/3rdparty/irodsphp/PHPUnitTest
 /apps/files_external/3rdparty/irodsphp/web
