+++ apps/webhook_listeners/appinfo/info.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0"?>
+<info xmlns:xsi= "http://www.w3.org/2001/XMLSchema-instance"
+	  xsi:noNamespaceSchemaLocation="https://apps.nextcloud.com/schema/apps/info.xsd">
+	<id>webhook_listeners</id>
+	<name>Nextcloud webhook support</name>
+	<summary>Nextcloud webhook support</summary>
+	<description>Nextcloud webhook support</description>
+	<version>1.0.0-dev</version>
+	<licence>agpl</licence>
+	<author>Côme Chilliet</author>
+	<namespace>WebhookListeners</namespace>
+
+	<types>
+		<filesystem/>
+	</types>
+
+	<category>customization</category>
+	<website>https://github.com/nextcloud/server</website>
+	<bugs>https://github.com/nextcloud/server/issues</bugs>
+	<repository>https://github.com/nextcloud/server.git</repository>
+
+	<dependencies>
+		<nextcloud min-version="30" max-version="30"/>
+	</dependencies>
+
+	<commands>
+		<command>OCA\WebhookListeners\Command\ListWebhooks</command>
+	</commands>
+
+	<settings>
+		<admin>OCA\WebhookListeners\Settings\Admin</admin>
+	</settings>
+</info>
