+++ apps/webhook_listeners/composer/composer.json
@@ -1,0 +1,13 @@
+{
+    "config" : {
+        "vendor-dir": ".",
+        "optimize-autoloader": true,
+        "classmap-authoritative": true,
+        "autoloader-suffix": "WebhookListeners"
+    },
+    "autoload" : {
+        "psr-4": {
+            "OCA\\WebhookListeners\\": "../lib/"
+        }
+    }
+}
