+++ apps/webhook_listeners/composer/composer/autoload_psr4.php
@@ -1,0 +1,10 @@
+<?php
+
+// autoload_psr4.php @generated by Composer
+
+$vendorDir = dirname(__DIR__);
+$baseDir = $vendorDir;
+
+return array(
+    'OCA\\WebhookListeners\\' => array($baseDir . '/../lib'),
+);
