--- apps/webhooks/appinfo/info.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0"?>
-<info xmlns:xsi= "http://www.w3.org/2001/XMLSchema-instance"
-	  xsi:noNamespaceSchemaLocation="https://apps.nextcloud.com/schema/apps/info.xsd">
-	<id>webhooks</id>
-	<name>Nextcloud webhook support</name>
-	<summary>Nextcloud webhook support</summary>
-	<description>Nextcloud webhook support</description>
-	<version>1.0.0-dev</version>
-	<licence>agpl</licence>
-	<author>Côme Chilliet</author>
-	<namespace>Webhooks</namespace>
-
-	<types>
-		<filesystem/>
-	</types>
-
-	<category>customization</category>
-	<website>https://github.com/nextcloud/server</website>
-	<bugs>https://github.com/nextcloud/server/issues</bugs>
-	<repository>https://github.com/nextcloud/server.git</repository>
-
-	<dependencies>
-		<nextcloud min-version="30" max-version="30"/>
-	</dependencies>
-
-	<commands>
-		<command>OCA\Webhooks\Command\ListWebhooks</command>
-	</commands>
-
-	<settings>
-		<admin>OCA\Webhooks\Settings\Admin</admin>
-	</settings>
-</info>
