--- apps/webhooks/composer/composer.json
@@ -1,13 +1,0 @@
-{
-    "config" : {
-        "vendor-dir": ".",
-        "optimize-autoloader": true,
-        "classmap-authoritative": true,
-        "autoloader-suffix": "Webhooks"
-    },
-    "autoload" : {
-        "psr-4": {
-            "OCA\\Webhooks\\": "../lib/"
-        }
-    }
-}
