--- apps/webhooks/composer/composer/autoload_psr4.php
@@ -1,10 +1,0 @@
-<?php
-
-// autoload_psr4.php @generated by Composer
-
-$vendorDir = dirname(__DIR__);
-$baseDir = $vendorDir;
-
-return array(
-    'OCA\\Webhooks\\' => array($baseDir . '/../lib'),
-);
