--- apps/webhooks/lib/Db/AuthMethod.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/**
- * SPDX-FileCopyrightText: 2024 Nextcloud GmbH and Nextcloud contributors
- * SPDX-License-Identifier: AGPL-3.0-or-later
- */
-
-namespace OCA\Webhooks\Db;
-
-enum AuthMethod: string {
-	case None = 'none';
-	case Header = 'header';
-}
